MCOL-1070 Fix exists in view subquery
When ExeMgr processes a correlated exists filter for a subquery inside a
view the de-duplication check doesn't work. We sometimes check unique
with the view name, sometimes not. We don't need the view name here so
remove it if we don't have it.

Also push the view name in the subquery.