MCOL-1579 Remove chmod of /dev/shm
We appear to chmod /dev/shm as a workaround to a CentOS 7 bug that was
fixed last year which accidentally set it to 755. If a user has /dev/shm
locked down we should get them to fix it rather than modifying it
ourselves.

The code before this fixed changed a root install to 755 for /dev/shm
which instantly broke anything using mmap() with MAP_SHARED as an
unprivileged user.