MCOL-3702: postConfig replication error
Todd & Andrew ID'd the problem as being that the libjemalloc.so ->
(real lib) link doesn't always exist for whatever reason.  I hacked our
run script to try the 3 possibilities mentioned in the ticket in this order
1) libjemalloc.so
2) libjemalloc.so.1
3) libjemalloc.so.2

Failing that, it will fall back to an empty LD_PRELOAD, so libc
would be used for the allocator.

Fixes the immediate problem.  The lib_bin and server_id vars mentioned
in the ticket are untouched by this commit.