MCOL-3764 Fix RENAME TABLE
When ALTER TABLE ... ENGINE= is called the following happens:

1. A temporary table is created in the new engine with the real table
   name but a temporary file name supplied
2. A bulk insert is started from old to new table
3. The old table is dropped
4. The new table is renamed

For #1 we use the real table name instead of the temporary file name
(otherwise step #2 breaks), we were therefore trying to skip #4.
This broke regular RENAME TABLE commands.

With this patch we detect if the rename is for a temporary to real table
and skip it. Since this is the only instance where we support temporary
tables.

This patch also fixes issues with extracting table names from file
names and some other irrelevant stuff.

Longer term if we want to support temporary tables we need to store the
provided filename in our metadata since it could be different from table
name.