MCOL-3793 Fix UM ACTIVE / DEGRADED flip-flop
OAM mysqld status check was recording the pid of mysqld but this was
being masked by another variable. So on every ACTIVE check it would
flip to DEGRADED and on every DEGRADED check it would flip to ACTIVE.

Variable masking is now removed. Also fixed a typo in the systemd check.