MCOL-3829: CS not starting correctly if rebooted at specific point
Found that the IDBDataFile path in BRM journal writing code needs to seek
to the end of the file before writing.

If save_brm is run (as it is during init), it truncates the journal file,
but the workernode retained its original offset, resulting in the front
of the file being 0-filled on the next journal write.  Load_brm can't
load that.