Took out the cache repopulate idea.  Now we will only have warnings of problems.
I realized we can't reliably tell how big the cache is while the system
is running.  There's a window where write/append has added / is adding
a journal file but hasn't told Cache about it yet.  This capability will
have to wait for now.

This shouldn't be a problem because in theory, we will no longer have
data whose size is not consistent with metadata stored outside of the file.
If we do, it means there was either a hard failure, or SM was killed.
Either way, SM will be restarted and the cache will populate its meta
fresh then.