Deb: Define Breaks/Replaces relationship on package libthrift-dev
The package libthrift-dev contains a file with the same name and
path as a file in mariadb-plugin-columnstore. The two packages
cannot be installed at the same time and by defining this
relationship in the debian/control file, apt will handle the case
automatically.

Error detected with check_for_missing_breaks_replaces.py:
  [ERROR] mariadb-plugin-columnstore conflicts with libthrift-dev
  files: {'/usr/lib/x86_64-linux-gnu/libthrift.so'}

Also ran 'wrap-and-sort -av' to ensure all debian/* contents is in
a deterministic order (=alphabetic).