MCOL-4738 AVG() returns a wrong result.
Release/RelWithDebInfo build on some platforms (Ubuntu18/20, Fedora)
is returning wrong results for the AVG() distributed function.

In Row::setLongDoubleField(), we remove the code that masks higher
bits of a long double field (if sizeof(long double) == 16 on a
given platform) since this masking appears to be redundant and also
fixes the issue with the AVG() function.