MCOL-4957 Fix performance slowdown for processing TIMESTAMP columns.
As part of MCOL-3776 to address synchronization issue while accessing
the fTimeZone member of the Func class, mutex locks were added to the
accessor and mutator methods. However, this slows down processing
of TIMESTAMP columns in PrimProc significantly as all threads across
all concurrently running queries would serialize on the mutex. This
is because PrimProc only has a single global object for the functor
class (class derived from Func in utils/funcexp/functor.h) for a given
function name. To fix this problem:

  (1) We remove the fTimeZone as a member of the Func derived classes
  (hence removing the mutexes) and instead use the fOperationType
  member of the FunctionColumn class to propagate the timezone values
  down to the individual functor processing functions such as
  FunctionColumn::getStrVal(), FunctionColumn::getIntVal(), etc.

  (2) To achieve (1), a timezone member is added to the
  execplan::CalpontSystemCatalog::ColType class.